created alma challenge handler lambda
