+++ serverless.yml
@@ -12,7 +12,7 @@ provider:
 
 functions:
   hello:
-    handler: handler.hello
+    handler: src/challenge-handler/handler.handleChallengeEvent
     events:
       - http:
           path: hello
