+++ test/challenge-handler/events/challenge.json
@@ -1,0 +1,8 @@
+{
+  "requestContext": {
+    "httpMethod": "GET"
+  },
+  "queryStringParameters": {
+    "challenge": "this is a challenge"
+  }
+}
